feat: extract disk index data
