+++ src/schema/index_util.cc
@@ -201,5 +201,67 @@ std::string IndexUtil::GetIDStr(const ::openmldb::common::ColumnKey& column_key)
     return id_str;
 }
 
+bool TableIndexInfo::Init() {
+    for (int32_t i = 0; i < table_meta_.column_desc_size(); i++) {
+        column_idx_map_.emplace(table_meta_.column_desc(i).name(), i);
+    }
+    uint32_t base_size = table_meta_.column_desc_size();
+    for (int32_t i = 0; i < table_meta_.added_column_desc_size(); ++i) {
+        column_idx_map_.emplace(table_meta_.added_column_desc(i).name(), i + base_size);
+    }
+    std::set<uint32_t> index_col_set;
+    for (int32_t i = 0; i < table_meta_.column_key_size(); i++) {
+        const auto& ck = table_meta_.column_key(i);
+        if (ck.flag()) {
+            continue;
+        }
+        for (const auto& add_ck : add_indexs_) {
+            if (ck.index_name() == add_ck.index_name()) {
+                add_index_idx_vec_.push_back(i);
+                break;
+            }
+        }
+        std::vector<uint32_t> cols;
+        for (const auto& name : ck.col_name()) {
+            auto iter = column_idx_map_.find(name);
+            if (iter != column_idx_map_.end()) {
+                cols.push_back(iter->second);
+                index_col_set.insert(iter->second);
+            } else {
+                return false;
+            }
+        }
+        index_cols_map_.emplace(i, std::move(cols));
+    }
+    if (add_index_idx_vec_.size() != add_indexs_.size()) {
+        return false;
+    }
+    std::map<uint32_t, uint32_t> col_idx_map;
+    for (auto idx : index_col_set) {
+        col_idx_map.emplace(idx, all_index_cols_.size());
+        all_index_cols_.push_back(idx);
+    }
+    for (const auto& kv : index_cols_map_) {
+        std::vector<uint32_t> vec;
+        for (auto idx : kv.second) {
+            vec.push_back(col_idx_map[idx]);
+        }
+        real_index_cols_map_.emplace(kv.first, std::move(vec));
+    }
+    return true;
+}
+
+bool TableIndexInfo::HasIndex(uint32_t idx) const {
+    return index_cols_map_.find(idx) != index_cols_map_.end();
+}
+
+const std::vector<uint32_t>& TableIndexInfo::GetIndexCols(uint32_t idx) {
+    return index_cols_map_[idx];
+}
+
+const std::vector<uint32_t>& TableIndexInfo::GetRealIndexCols(uint32_t idx) {
+    return real_index_cols_map_[idx];
+}
+
 }  // namespace schema
 }  // namespace openmldb
