+++ src/storage/mem_table_snapshot.h
@@ -113,28 +113,6 @@ class DataReader {
     uint64_t failed_cnt_ = 0;
 };
 
-class TableIndexInfo {
- public:
-    TableIndexInfo(const ::openmldb::api::TableMeta& table_meta,
-             const std::vector<::openmldb::common::ColumnKey>& add_indexs)
-        : table_meta_(table_meta), add_indexs_(add_indexs) {}
-    bool Init();
-    const std::vector<uint32_t>& GetAllIndexCols() const { return all_index_cols_; }
-    const std::vector<uint32_t>& GetAddIndexIdx() const { return add_index_idx_vec_; }
-    bool HasIndex(uint32_t idx) const;
-    const std::vector<uint32_t>& GetIndexCols(uint32_t idx);
-    const std::vector<uint32_t>& GetRealIndexCols(uint32_t idx);  // the pos in all_index_cols_
-
- private:
-    ::openmldb::api::TableMeta table_meta_;
-    std::vector<::openmldb::common::ColumnKey> add_indexs_;
-    std::map<std::string, uint32_t> column_idx_map_;
-    std::vector<uint32_t> all_index_cols_;
-    std::vector<uint32_t> add_index_idx_vec_;
-    std::map<uint32_t, std::vector<uint32_t>> index_cols_map_;
-    std::map<uint32_t, std::vector<uint32_t>> real_index_cols_map_;
-};
-
 struct DeleteSpan {
     DeleteSpan() = default;
     explicit DeleteSpan(const api::LogEntry& entry);
@@ -188,7 +166,7 @@ class MemTableSnapshot : public Snapshot {
     base::Status ExtractIndexData(const std::shared_ptr<Table>& table,
             const std::vector<::openmldb::common::ColumnKey>& add_indexs,
             const std::vector<std::shared_ptr<::openmldb::log::WriteHandle>>& whs,
-            uint64_t offset, bool dump_data);
+            uint64_t offset, bool dump_data) override;
 
     int CheckDeleteAndUpdate(std::shared_ptr<Table> table, ::openmldb::api::LogEntry* new_entry);
 
@@ -200,9 +178,6 @@ class MemTableSnapshot : public Snapshot {
                                std::atomic<uint64_t>* g_failed_cnt);
 
     uint64_t CollectDeletedKey(uint64_t end_offset);
-
-    ::openmldb::base::Status DecodeData(const std::shared_ptr<Table>& table, const openmldb::api::LogEntry& entry,
-            const std::vector<uint32_t>& cols, std::vector<std::string>* row);
 
     std::string GenSnapshotName();
