+++ src/storage/snapshot.h
@@ -55,6 +55,9 @@ class Snapshot {
                                 ::openmldb::api::Manifest& manifest);  // NOLINT
     std::string GetSnapshotPath() { return snapshot_path_; }
 
+    ::openmldb::base::Status DecodeData(const std::shared_ptr<Table>& table, base::Slice raw_data,
+            const std::vector<uint32_t>& cols, std::vector<std::string>* row);
+
     virtual base::Status ExtractIndexData(const std::shared_ptr<Table>& table,
             const std::vector<::openmldb::common::ColumnKey>& add_indexs,
             const std::vector<std::shared_ptr<::openmldb::log::WriteHandle>>& whs,
